moved new test to ElasticSearchDataContextTest.java so that code would compile with RAT licence issues
